Add ResourceCache
Adds an official resource cache. It uses types of the resource it wants as well as the types of its arguments, combines that information into a key, and uses that to map to a resource. This cache uses the LRU eviction policy and keeps track of the "age" of each entry.

The goal for later  is to try and integrate this and remove the old SampleCache implemented here.