Implement "direct routing" optimization
This new optimization allows SampleFrame-based plugins (2-channel
interleaved) to bypass the pin connector routing and intermediate
buffers entirely when the pins are the set to the default configuration
(L-->L, R-->R).

This should make performance for plugins that have a pin connector about
the same as before pin connector support was added for the common-case
default pin configuration. In other words, users will not need to pay
for what they don't use.

In order to implement this optimization, I moved the
"provideProcessBuffers" compile-time option to the `AudioPortsConfig`
struct and renamed it "buffered". I also added a `process` method to the
`AudioPortsModel::Router` class which allowed me to simplify some of the
AudioPlugin code. Then I simplified it a bit further using a new
`AudioDataViewType` metafunction.

Lastly, I added a unit test called "DirectRoutingOptimization" which
verifies that AudioPortsModel correctly recognizes when the direct
routing optimization can be applied.