Remove `BufferManager`
For a while now BufferManager hasn't done anything except allocate and
deallocate a buffer using the global new/delete operators. This is not
enough to justify its existence, and since it was proving to be
problematic now with the usage of `std::span<SampleFrame>` over
`SampleFrame*`, I decided to just remove it.