Dont try to resample within the effect chain for LADSPA
LADSPA was resampling inside the effect chain,
which is improper as input data is needed at
the ends of the buffer in order to produce the correct resampling output, which cant happen.

Instead, add a TODO comment that says we should add a disclaimer that using certain sample rates may cause issues with some plugins. Most users
are going to be using either 44.1khz or 48khz, and they should be aware that there may be problems using other sample rates that some plugins dont work well with.