Allocate internal callback buffer on the heap
I would prefer allocating on the stack here,
but its safer to allocate on the heap because
of the variable channel count.

The allocation only happens once, so its
not a big performance determinent (both
stack memory and heap memory are in RAM,
even though stack memory could be easier to work with in general from a hardware perspective).