Incomplete input selection via combo box
Add two combo boxes and a Jack client to the Jack setup dialog. The combo
boxes display the available Jack inputs. Please note that in Jack
terminology these are called "outputs".

The own Jack client is necessary because the setup dialog does not have
any access to the actual driver. So a new client is created when the
dialog is opened and deleted when it is closed, i.e. when the dialog is
deleted itself.

The available inputs are collected via
`AudioJack::setupWidget::getAudioInputNames` and are then put into the
combo boxes via `AudioJack::setupWidget::populateComboBox`.

`AudioJack::setupWidget::saveSettings` saves the selections that have
been made in the combo boxes into the configuration.

Add the method `handleRegistrationEvent` to `AudioJack` and register it
in `AudioJack::initJackClient` via `jack_set_port_registration_callback`.
Currently it will only output information about the port via `printf` but
not do anything else. It can likely be removed again because the
`AudioJack::setupWidget` uses it's own client and must register its own
callbacks in case it wants to react to changed inputs and outputs while
the setup dialog is open.

* Do the same for the inputs
* Read the information stored in the configuration when the driver is
  initialized. Gracefully handle when selections are not available, e.g.
  because the device is not plugged in.
* Decide if the setup dialog should react to changed ports while it is
  open. This might complicate selections that are currently made, i.e. it
  has to be ensured that the user experience is good.
* Remove printf statements