Add combo boxes for Jack outputs
Extend the Jack setup dialog with combo boxes that present the available
outputs.

Save the selected outputs in the configuration.

Add `AudioJack::setupWidget::getAudioPortNames` which takes the type of
port and then collects all port names which match. Make the new
`getAudioOutputNames` and `getAudioInputNames` delegate to that method
with the appropriate type. This also hides the different terminologies a
bit.