Show warning message for missing ports
Show a warning message in the Jack setup dialog if some inputs or outputs
that are set in the configuration are missing in the system. This might
for example be caused by disconnected devices.

Let `populateComboBox` return a boolean which indicates if the selected
port can be found in the list of displayed ports or not.