Combine overloads of fastRand()
Combines the overloads for floating point types and integral types by
making the FAST_RAND_RATIO type conditional. It is now the same type as
the input range if the range is a floating point type, otherwise it
defaults to float. This removes the need for the FAST_RAND_MAX constant.