Misc. changes
- Move Effect's RMS out sum calculation to checkGate(). Now the
calculation is only performed when auto-quit is enabled.
- Rename `Effect::processAudioBuffer[Impl]` to
`Effect::processCore[Impl]` and use `InterleavedBufferView`. The "Core"
makes it clear that LMMS core calls it with the track channels. In a
follow-up PR, the name for the process method used by instruments will
also be named `processCore` for consistency.
- New compile-time validation check for `AudioPortsSettings`
- Updated a reference to `processAudioBuffer` in the SpectrumAnalyzer
README
- Updated names for some unit tests now that `SampleFrame` is no longer
used