Fix release builds by unconditional output
Make the output unconditional again because it interfered with some
release builds. They complained that the `result` variable in
`AudioJack::attemptToConnect` was unused because it was only further used
in debug builds.

The alternative would have been to duplicate the `jack_connect` call.
However, that would have been a worse solution because it would have
duplicated critical code which might have deviated due to the
duplication.