Fix RMS calculation for effects
`AudioPortsModel` now keeps track of which processor output channels are
used (`AudioPortsModel::m_usedProcessorChannels`), and the RMS is
calculated from the buffers of only those channels.

`AudioPorts::Router` is responsible for calculating the RMS when needed.
If `router.process()` returns `ProcessStatus::ContinueIfNotQuiet`, the
RMS can be obtained through `router.rms()`.

This required refactoring the router class a little, but it no longer
requires a partial specialization and the process method returns
`ProcessStatus` now, which is much nicer.

I also improved some of the cache update and channel count modification
code in `AudioPortsModel` in order to support `m_usedProcessorChannels`
and reduce the amount of cache update calculations as much as possible.
I added more tests for all of this.

TODO: I should probably zero any track channel buffers a processor
routes to if it is asleep due to RMS == 0. To do this efficiently
though, I'll need to keep track of which track channels are already
zeroed and only zero the buffers when they need to be.