Improve AudioPorts initialization and activation
- Add `initialized` method to `AudioPorts::Buffer` interface
- Make `AudioPorts::active()` mean `AudioPortsModel` and the buffers are
fully initialized and ready to be used by the audio processor's process
method (`processImpl`)
- Make `AudioPorts::buffers()` simply return the buffers if it is able
to
- Rename `bufferPropertiesChanged` virtual method to
`bufferPropertiesChanging` in `AudioPortsModel`. This allows
implementations of the virtual method to to check what the old channel
counts were, which may be useful.
- Make the cached data in `AudioPortsModel` private and add protected
getters
- Simplify the `RemotePluginAudioPorts` implementation now that "active"
and "initialized" are well-defined
- Fix minor bug where audio ports for VST plugins were activated before
the RemotePlugin was fully initialized
- Simplify `AudioPortsModel::Matrix::setTrackChannelCount` and
`AudioPortsModel::Matrix::setChannelCount`