Remove quality settings (interpolation mode)
The quality settings can conflict with an instrument's choice of interpolation. For example, one instrument may only work with a fixed set of interpolation modes that do not necessarily align with the interpolation modes specified by AudioResampler. AudioRessampler works with some modes like SincBest, but some instruments may not have that mode, so theres a conflict. It is simpler and safer
to allow instruments to determine what interpolation
mode they will use.