Keep track of track channel silence + auto-quit rework
The AudioBus now keeps track of which of its track channels are known to
be silent. This information can be used to avoid zeroing buffers
unnecessarily.

But more importantly, it also allowed me to rework the auto-quit system
for effects. With this commit, sleeping effects determine whether they
need to wake up by simply checking if there are any non-quiet track
channels routed into their inputs. This is an O(1) task now that
AudioBus keeps track of that info. The previous auto-quit system was
hacky and made every effect in the chain wake up at once, even if there
was no input noise routed to one of their inputs. With the new system,
effects can wake up independently of each other and only when needed.

- AudioBus moved to its own .h/.cpp files
- Added a bitset to AudioBus to store which track channels are currently
known to be quiet
- Added AudioBus methods for checking for input noise, sanitizing
buffers, clearing buffers, and updating silence flags
- Integrated the new AudioBus functionality into AudioPlugin,
AudioPorts::Router, AudioBusHandle, EffectChain, Mixer, and other places
where needed in order for the silence flags of track channels to
propagate and update correctly
- Moved some AudioPorts::Router functionality to AudioBus
- Make Effect::startRunning and Effect::stopRunning protected now that
they are not used outside the effect
- Updated the unit tests