Add small patch for PortAudio
LMMS doesnt stop PortAudio streams correctly as it doesnt correctly distinguish between "active" and "stopped" PortAudio streams. In LMMS, stopping the audio engine can cause the PortAudio stream to be in a state of inactivity, but not stopped. Stopping the PortAudio device in the code only calls Pa_StopStream if the stream is active, but that doesn't mean the stream is stopped. This caused partial stopping behavior and needed to be partially fixed here or else the whole application freezes when using PortAudio. This is a brief description as the details are more lengthy in nature.

In reality, this check for if the stream is stopped is arguably redundant, and its more safer to just forward the start/stop behavior to Pa_StartStream and Pa_StopStream directly without any conditional logic. There is a larger refactoring on the PortAudio backend in another PR that does this, so this fix is temporary.